Move export function
