+++ src/Web/ClientApp/src/interfaces/controller/IHomeController.ts
@@ -1,1 +1,6 @@
-export interface IHomeController {}
+import { InspectionSheet } from "../../entities";
+
+export interface IHomeController {
+  exportExcelInspectionSheet(sheet: InspectionSheet): void;
+  exportJsonInspectionSheet(sheet: InspectionSheet): void;
+}
